backend/back/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.4.1
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.0
hexbytes==1.3.1
idna==3.10
multidict==6.4.3
parsimonious==0.10.0
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyunormalize==16.0.0
pywin32==310
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
sqlparse==0.5.3
toolz==1.0.0
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
web3==7.11.1
websockets==15.0.1
yarl==1.20.0
